+++ services/api/src/handlers.rs
@@ -1,0 +1,1 @@
+pub mod users;
