+++ services/api/src/handlers/users.rs
@@ -1,0 +1,1 @@
+pub mod create_user;
